chore: Addressed Devin's concern: "Duplicated math-segment parsing logic between latex.rs and parser.rs"
