+++ AGENTS.md
@@ -53,6 +53,7 @@ The core code is structured into the following modules within the `src` director
 - `cli.rs`: Contains the command-line interface definitions using Clap.
 - `config.rs`: Contains runtime configuration validation and filter handling for CLI arguments.
 - `latex.rs`: Contains logic for handling LaTeX-specific formatting and LaTeX substitutions.
+- `math.rs`: Contains logic for handling LaTeX mathematical expressions, including splitting math segments and checking for escaped characters.
 - `lib.rs`: Exposes the library modules used by the binary and tests.
 
 Outside the `src` directory, we have:
