+++ src/latex.rs
@@ -8,6 +8,8 @@ use std::path::Path;
 
 use anyhow::{Context, Result};
 
+use crate::math::{is_escaped, split_math_segments};
+
 /// Deterministic substitution map type used by the application.
 pub type SubstitutionMap = BTreeMap<String, String>;
 
@@ -93,12 +95,11 @@ pub fn substitute_latex_to_text_with_ordered(
     let mut output = String::new();
 
     for segment in segments {
-        match segment {
-            Segment::Text(text) => {
-                let unwrapped = unwrap_formatting_commands(&text);
-                output.push_str(&apply_substitutions(&unwrapped, ordered_substitutions));
-            }
-            Segment::Math(math) => output.push_str(&math),
+        if segment.is_math {
+            output.push_str(&segment.text);
+        } else {
+            let unwrapped = unwrap_formatting_commands(&segment.text);
+            output.push_str(&apply_substitutions(&unwrapped, ordered_substitutions));
         }
     }
 
@@ -136,96 +137,6 @@ fn apply_substitutions(input: &str, ordered_substitutions: &[(String, String)]) 
     }
 
     output
-}
-
-#[derive(Debug)]
-enum Segment {
-    Text(String),
-    Math(String),
-}
-
-fn split_math_segments(input: &str) -> Vec<Segment> {
-    let chars = input.chars().collect::<Vec<_>>();
-    let mut index = 0;
-    let mut text_buffer = String::new();
-    let mut segments = Vec::new();
-
-    while index < chars.len() {
-        if let Some((math_segment, next_index)) = consume_math_segment(&chars, index) {
-            if !text_buffer.is_empty() {
-                segments.push(Segment::Text(std::mem::take(&mut text_buffer)));
-            }
-            segments.push(Segment::Math(math_segment));
-            index = next_index;
-            continue;
-        }
-
-        text_buffer.push(chars[index]);
-        // When an unclosed `$$` is detected, consume both `$` characters as plain
-        // text so the second `$` is not misinterpreted as a single-`$` delimiter.
-        if chars[index] == '$'
-            && !is_escaped(&chars, index)
-            && index + 1 < chars.len()
-            && chars[index + 1] == '$'
-        {
-            text_buffer.push(chars[index + 1]);
-            index += 2;
-        } else {
-            index += 1;
-        }
-    }
-
-    if !text_buffer.is_empty() {
-        segments.push(Segment::Text(text_buffer));
-    }
-
-    segments
-}
-
-fn consume_math_segment(chars: &[char], start: usize) -> Option<(String, usize)> {
-    if start >= chars.len() {
-        return None;
-    }
-
-    if chars[start] == '$' && !is_escaped(chars, start) {
-        if start + 1 < chars.len() && chars[start + 1] == '$' {
-            return extract_delimited_segment(chars, start, "$$", 2);
-        }
-
-        return extract_delimited_segment(chars, start, "$", 1);
-    }
-
-    if chars[start] == '\\' && !is_escaped(chars, start) && start + 1 < chars.len() {
-        return match chars[start + 1] {
-            '(' => extract_delimited_segment(chars, start, "\\)", 2),
-            '[' => extract_delimited_segment(chars, start, "\\]", 2),
-            _ => None,
-        };
-    }
-
-    None
-}
-
-fn extract_delimited_segment(
-    chars: &[char],
-    start: usize,
-    close: &str,
-    open_len: usize,
-) -> Option<(String, usize)> {
-    let close_chars = close.chars().collect::<Vec<_>>();
-    let close_len = close_chars.len();
-    let mut index = start + open_len;
-
-    while index + close_len <= chars.len() {
-        let is_match = chars[index..index + close_len] == close_chars[..];
-        if is_match && !is_escaped(chars, index) {
-            let content = chars[start..index + close_len].iter().collect::<String>();
-            return Some((content, index + close_len));
-        }
-
-        index += 1;
-    }
-    None
 }
 
 fn unwrap_formatting_commands(input: &str) -> String {
@@ -336,25 +247,6 @@ fn extract_braced_content(chars: &[char], open_brace_index: usize) -> Option<(St
     }
 
     None
-}
-
-fn is_escaped(chars: &[char], index: usize) -> bool {
-    if index == 0 {
-        return false;
-    }
-
-    let mut slash_count = 0;
-    let mut lookback = index;
-    while lookback > 0 {
-        lookback -= 1;
-        if chars[lookback] == '\\' {
-            slash_count += 1;
-        } else {
-            break;
-        }
-    }
-
-    slash_count % 2 == 1
 }
 
 #[cfg(test)]
