+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod cli;
 pub mod config;
 pub mod latex;
+pub(crate) mod math;
 pub mod parser;
 pub mod template;
 pub mod utils;
