+++ src/parser.rs
@@ -12,6 +12,8 @@ use biblatex::{Bibliography, Chunk, ChunksExt, RawBibliography, RawChunk};
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
+use crate::math::split_math_segments;
+
 /// Error types for BibTeX parsing operations
 #[derive(Error, Debug)]
 pub enum ParseError {
@@ -488,8 +490,8 @@ impl BibTeXParser {
     }
 
     fn merge_original_math_segments(formatted: &str, original: &str) -> String {
-        let formatted_segments = Self::split_math_segments(formatted);
-        let original_math_segments = Self::split_math_segments(original)
+        let formatted_segments = split_math_segments(formatted);
+        let original_math_segments = split_math_segments(original)
             .into_iter()
             .filter(|segment| segment.is_math)
             .map(|segment| segment.text)
@@ -533,120 +535,6 @@ impl BibTeXParser {
         }
 
         output
-    }
-
-    fn split_math_segments(value: &str) -> Vec<MathAwareSegment> {
-        let chars = value.chars().collect::<Vec<_>>();
-        let mut index = 0;
-        let mut text_buffer = String::new();
-        let mut segments = Vec::new();
-
-        while index < chars.len() {
-            if let Some((math_segment, next_index)) = Self::consume_math_segment(&chars, index) {
-                if !text_buffer.is_empty() {
-                    segments.push(MathAwareSegment {
-                        is_math: false,
-                        text: std::mem::take(&mut text_buffer),
-                    });
-                }
-
-                segments.push(MathAwareSegment {
-                    is_math: true,
-                    text: math_segment,
-                });
-                index = next_index;
-                continue;
-            }
-
-            text_buffer.push(chars[index]);
-            // When an unclosed `$$` is detected, consume both `$` characters as plain
-            // text so the second `$` is not misinterpreted as a single-`$` delimiter.
-            if chars[index] == '$'
-                && !Self::is_escaped(&chars, index)
-                && index + 1 < chars.len()
-                && chars[index + 1] == '$'
-            {
-                text_buffer.push(chars[index + 1]);
-                index += 2;
-            } else {
-                index += 1;
-            }
-        }
-
-        if !text_buffer.is_empty() {
-            segments.push(MathAwareSegment {
-                is_math: false,
-                text: text_buffer,
-            });
-        }
-
-        segments
-    }
-
-    fn consume_math_segment(chars: &[char], start: usize) -> Option<(String, usize)> {
-        if start >= chars.len() {
-            return None;
-        }
-
-        if chars[start] == '$' && !Self::is_escaped(chars, start) {
-            if start + 1 < chars.len() && chars[start + 1] == '$' {
-                return Self::extract_delimited_segment(chars, start, "$$", 2);
-            }
-
-            return Self::extract_delimited_segment(chars, start, "$", 1);
-        }
-
-        if chars[start] == '\\' && !Self::is_escaped(chars, start) && start + 1 < chars.len() {
-            return match chars[start + 1] {
-                '(' => Self::extract_delimited_segment(chars, start, "\\)", 2),
-                '[' => Self::extract_delimited_segment(chars, start, "\\]", 2),
-                _ => None,
-            };
-        }
-
-        None
-    }
-
-    fn extract_delimited_segment(
-        chars: &[char],
-        start: usize,
-        close: &str,
-        open_len: usize,
-    ) -> Option<(String, usize)> {
-        let close_chars = close.chars().collect::<Vec<_>>();
-        let close_len = close_chars.len();
-        let mut index = start + open_len;
-
-        while index + close_len <= chars.len() {
-            let is_match = chars[index..index + close_len] == close_chars[..];
-            if is_match && !Self::is_escaped(chars, index) {
-                let segment = chars[start..index + close_len].iter().collect::<String>();
-                return Some((segment, index + close_len));
-            }
-
-            index += 1;
-        }
-
-        None
-    }
-
-    fn is_escaped(chars: &[char], index: usize) -> bool {
-        if index == 0 {
-            return false;
-        }
-
-        let mut slash_count = 0;
-        let mut lookback = index;
-        while lookback > 0 {
-            lookback -= 1;
-            if chars[lookback] == '\\' {
-                slash_count += 1;
-            } else {
-                break;
-            }
-        }
-
-        slash_count % 2 == 1
     }
 
     fn parse_slugified_keywords(value: &str) -> Vec<String> {
@@ -724,11 +612,6 @@ impl BibTeXParser {
 
         AuthorName { first, last, full }
     }
-}
-
-struct MathAwareSegment {
-    is_math: bool,
-    text: String,
 }
 
 #[cfg(test)]
